requirements.txt
django-allauth
django-allauth[socialaccount]